+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/TxAuxData.hs
@@ -13,6 +13,7 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE UndecidableSuperClasses #-}
 {-# LANGUAGE ViewPatterns #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -20,6 +21,8 @@ module Cardano.Ledger.Allegra.TxAuxData (
   AllegraTxAuxData (AllegraTxAuxData),
   AllegraTxAuxDataRaw,
   metadataAllegraTxAuxDataL,
+  AllegraEraTxAuxData (..),
+  timelockScriptsAllegraTxAuxDataL,
 
   -- * Deprecations
   AuxiliaryData,
@@ -90,6 +93,9 @@ data AllegraTxAuxDataRaw era = AllegraTxAuxDataRaw
   }
   deriving (Generic, Eq)
 
+class EraTxAuxData era => AllegraEraTxAuxData era where
+  timelockScriptsTxAuxDataL :: Lens' (TxAuxData era) (StrictSeq (Timelock era))
+
 instance Crypto c => EraTxAuxData (AllegraEra c) where
   type TxAuxData (AllegraEra c) = AllegraTxAuxData (AllegraEra c)
 
@@ -106,6 +112,14 @@ instance Crypto c => EraTxAuxData (AllegraEra c) where
 metadataAllegraTxAuxDataL :: Era era => Lens' (AllegraTxAuxData era) (Map Word64 Metadatum)
 metadataAllegraTxAuxDataL =
   lensMemoRawType atadrMetadata $ \txAuxDataRaw md -> txAuxDataRaw {atadrMetadata = md}
+
+instance Crypto c => AllegraEraTxAuxData (AllegraEra c) where
+  timelockScriptsTxAuxDataL = timelockScriptsAllegraTxAuxDataL
+
+timelockScriptsAllegraTxAuxDataL ::
+  Era era => Lens' (AllegraTxAuxData era) (StrictSeq (Timelock era))
+timelockScriptsAllegraTxAuxDataL =
+  lensMemoRawType atadrTimelock $ \txAuxDataRaw ts -> txAuxDataRaw {atadrTimelock = ts}
 
 deriving instance HashAlgorithm (HASH (EraCrypto era)) => Show (AllegraTxAuxDataRaw era)