+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/TxAuxData.hs
@@ -3,16 +3,17 @@
 
 module Cardano.Ledger.Babbage.TxAuxData () where
 
+import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.TxAuxData (
   AlonzoTxAuxData (..),
   hashAlonzoTxAuxData,
   metadataAlonzoTxAuxDataL,
+  timelockScriptsAlonzoTxAuxDataL,
   translateAlonzoTxAuxData,
   validateAlonzoTxAuxData,
  )
 import Cardano.Ledger.Babbage.Era
 import Cardano.Ledger.Babbage.Scripts ()
-import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto
 
 instance Crypto c => EraTxAuxData (BabbageEra c) where
@@ -27,3 +28,6 @@ instance Crypto c => EraTxAuxData (BabbageEra c) where
   hashTxAuxData = hashAlonzoTxAuxData
 
   validateTxAuxData = validateAlonzoTxAuxData
+
+instance Crypto c => AllegraEraTxAuxData (BabbageEra c) where
+  timelockScriptsTxAuxDataL = timelockScriptsAlonzoTxAuxDataL