+++ eras/allegra/impl/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add a `ToJSON` instance for `ValidityInterval`
 * Add `metadataAllegraTxAuxDataL`
+* Add `AllegraEraTxAuxData` with `timelockScriptsTxAuxDataL` and helper
+  `timelockScriptsAllegraTxAuxDataL`
 
 ## 1.4.0.0